pay/requirements.txt

asgiref==3.5.0
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==3.2
django-admin-rangefilter==0.9.0
django-allauth==0.51.0
django-taggit==3.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
phonenumbers==8.13.11
Pillow==9.1.1
pipenv==2022.12.19
PyJWT==1.7.1
PyMySQL==1.0.3
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2


